Update CRUD opreration on products to load images into bucket

+++ app/actions/delete-product.ts
@@ -2,14 +2,26 @@ import { supabase } from "@/lib/supabase/client";
 
 const defaultErrorMessage = 'Failed to delete product';
 
-export const deleteProduct = async (id: string) => {
+export const deleteProduct = async (product: ProductItem) => {
   try {
+
+    const { error: errorImage } = await supabase
+      .storage
+      .from('product-images')
+      .remove([product.image]);
+
+    if (errorImage ) {
+      return {
+        success: false,
+        error: errorImage?.message ?? defaultErrorMessage
+      };
+    };
     const { error } = await supabase
       .from('products')
       .delete()
-      .eq('id', id);
+      .eq('id', product.id);
 
-    if (error) {
+    if (error ) {
       return {
         success: false,
         error: error?.message ?? defaultErrorMessage
